Code review round 3 tier 2 - Drop unused `ref` param from dynamic_conv_tolerances / _winograd_tolerances; update call sites. - Move bench-only flops_conv to bench_conv2d.py; test-only apply_activation + tolerance fns to _helpers.py (bench imports them). - _utils.py now holds production helpers only (_storage_ptr stays — used by _prepack.py).
